requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
Pillow==6.1.0
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
urllib3==1.25.3
